Fix vulnerability issue
Hi Adam! I realised our code was potentially open to Cross-Site Scripting threats.
I worked on it and instead of creating a new note using just a template literal, the title and body are now generated separately and the user's input is passed in as text using the textContent property to prevent attacks 👾 👍🏻
Just in case we want to use the app with a database..
Let me know if I made mistakes and if it makes sense to you too! 🙌🏻